requirements.txt
tqdm
keras
numpy
scipy
docker
grpcio
pandas
seaborn
matplotlib
tensorflow
grpcio-tools
opencv-python
